requirements.txt
astroid==2.2.5
attrs==19.1.0
axel==0.0.7
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
DateTimeRange==0.5.5
futures-then==0.1.1
idna==2.8
iso8601==0.1.12
isort==4.3.17
joblib==0.13.2
jsons==0.8.6
kafka-python==1.4.3
lazy-object-proxy==1.3.1
marshmallow==3.0.0rc5
marshmallow-enum==1.4.1
mbstrdecoder==0.7.0
mccabe==0.6.1
numpy==1.16.2
pandas==0.23.4
pylint==2.3.1
python-dateutil==2.8.0
requests==2.21.0
serpy==0.3.1
six==1.12.0
sortedcontainers==2.1.0
typed-ast==1.3.4
typepy==0.4.0
urllib3==1.24.1
wrapt==1.11.1
mat.ocs.streaming==1.1.0+58.g1e303b3